pyproject.toml
[project]
name = "nanobot-deep"
version = "0.1.0"
description = "LangGraph/DeepAgents integration for nanobot"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nanobot-deep contributors"}
]
keywords = ["ai", "agent", "langgraph", "deepagents"]
dependencies = [
    "nanobot-ai>=0.1.4.post4,<0.2.0",
    "deepagents>=0.4.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.1.0",
    "httpx>=0.25.0",
]
observability = [
    "langfuse>=2.0.0",
]

[project.scripts]
nanobot-deep = "nanobot_deep.cli:run_nanobot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot_deep"]

[tool.hatch.build.targets.wheel.sources]
"nanobot_deep" = "nanobot_deep"

[tool.hatch.build]
include = [
    "nanobot_deep/**/*.py",
    "templates/**/*.json",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
